requirements.txt
aioredis==1.3.1
amqp==2.6.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.6
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-timezone-field==4.0
djangorestframework==3.11.0
future==0.18.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
jsonschema==3.2.0
kombu==4.6.11
msgpack==1.0.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
vine==1.3.0
zope.interface==5.1.0
